Fixed syntax errors, need to fix runtime errors

+++ src/main/scala/main.scala
@@ -13,7 +13,7 @@ object SIFT {
 
         chiselMainTest(tutArgs,
           () => Module(new ScaleSpaceExtrema(
-            new ImageType(UInt(img.w), UInt(img.h), img.d)))) {
+            new ImageType(img.w, img.h, img.d)))) {
             c => new ScaleSpaceExtremaTests(c, "data/in.im24",
               "data/out.im24","data/coord.im24")
       }
